requirements.txt
scikit_image==0.14.2
torchvision==0.2.2.post3
numpy==1.16.2
torch==0.4.1
Pillow==6.0.0
graphviz==0.11
skimage==0.0
